+++ .gitignore
@@ -1,5 +1,16 @@
-personal_assistant_bot/pycache
-personal_assistant_bot/classes/pycache
-personal_assistant_bot/classes/__pycache__
-personal_assistant_bot/__pycache__
+personal_assistant_bot/pycache/
+personal_assistant_bot/classes/pycache/
+personal_assistant_bot/settings/pycache/
+personal_assistant_bot/functions/pycache/
+personal_assistant_bot/classes/__pycache__/
+personal_assistant_bot/settings/__pycache__/
+personal_assistant_bot/functions/__pycache__/
+personal_assistant_bot/__pycache__/
+
+clear/
+test/
 bot/
+bo/
+
+.gitignore
+
