+++ personal_assistant_bot/classes/address.py
@@ -1,4 +1,7 @@
-from .field import Field
+import sys
+sys.path.append('..')
+
+from classes.field import Field
 
 
 class Address(Field):
