+++ personal_assistant_bot/classes/email.py
@@ -1,6 +1,9 @@
 import re
 
-from .field import Field
+import sys
+sys.path.append('..')
+
+from classes.field import Field
 
 from colorama import init, Fore
 init(autoreset=True)
