--- personal_assistant_bot/classes/settings.py
@@ -1,6 +1,0 @@
-from pathlib import Path
-
-filename = Path(Path.home(), "adressbook.pkl")
-PAG = 4
-
-notes_filename = Path(Path.home(), "notes.pkl")
