+++ personal_assistant_bot/settings/settings.py
@@ -1,0 +1,8 @@
+from pathlib import Path
+
+addressbook_filename = Path(Path.home(), "adressbook.pkl")
+notes_filename = Path(Path.home(), "notes.pkl")
+
+PAG = 4
+NOTE_LEN = 40
+
